Move functionality contained in the split(), glyphs() and widths() methods of SingleStyledText to Paragraph.render()
This avoids creation of SingleStyledText instances for each word in a
SingleStyledText, and thus the initialization in widths() and glyphs() does
not need to be performed over and over again.

The RFIC example still renders, but the output is broken. Also, some
features such as tabs are currently being ignored.