Fix GroupedLabeledFlowables label width
If a label exceeded label_max_width, the content of each of the labeled
flowables would be placed at label_max_width + label_spacing. If the
content is to be placed below the label, it might as well be indented
less than max_label_width. Either the width of the widest label smaller
than max_label_width (+ spacing), or min_label_width.