Sphinx style sheet: Improve definition list styling
Setting label_max_width < label_min_width causes definitions with a
term exceeding label_min_width to be rendered on a separate line.

Definitions used to be rendered like this, but recent changes must
have changed this. This change was not detected because we no longer
have tests in place to detect this regression (see #169).