Disregard descenders near borders (fixes #144)
When a flowable has a top border, the descender of the previous flowable
is now being ignored. Similarly, when there is a bottom border, the
flowable's descender is not being passed on to the next.

The reasoning here is that borders separate flowables so that descenders
should not be taken into account.