src\rinoh\stylesheets\matcher.py wildcard-import
Code uses from rinoh.styleds import *
This hides the imoported objects and might lead to collisions in the future if objects of the same name are imported with wildcards.

styleds is a collection of imports, aimed to ease related imports.
I copied the used imports from there and deleted the unused ones.

This adds many imports and does not  benefits from styleds, so this is a downside of fixing this alert.